Add --skip_markov and --depth 0 to sample_markov.
Additionally, make Quantizable return u32s

+++ src/quantize.rs
@@ -10,7 +10,7 @@ pub trait Quantizable: Sized {
     fn from_f64(x: f64) -> Self;
     fn into_f64(x: Self) -> f64;
 
-    fn unquantize(x: usize, min: Self, max: Self, quantization_level: usize) -> Self {
+    fn unquantize(x: u32, min: Self, max: Self, quantization_level: u32) -> Self {
         let min = Self::into_f64(min);
         let max = Self::into_f64(max);
 
@@ -23,12 +23,12 @@ pub trait Quantizable: Sized {
         Self::from_f64(x)
     }
 
-    /// Quantize `x` to a usize. The `min` and `max` parameters should be set to the
+    /// Quantize `x` to a u32. The `min` and `max` parameters should be set to the
     /// largest and smallest values that `x` may take on. The `quantization_level` parameter
-    /// determines the range of the output usize, and consequently, the number of values
+    /// determines the range of the output u32, and consequently, the number of values
     /// the quantization can take on. For example, a quantization_level of 256 means
     /// that the number of representable values is reduced to 256.
-    fn quantize(x: Self, min: Self, max: Self, quantization_level: usize) -> usize {
+    fn quantize(x: Self, min: Self, max: Self, quantization_level: u32) -> u32 {
         let min = Self::into_f64(min);
         let max = Self::into_f64(max);
         let x = Self::into_f64(x);
@@ -36,9 +36,9 @@ pub trait Quantizable: Sized {
         // First, map x from the [min, max] range to the [0.0, 1.0] range (which turns x into t)
         let t = inv_lerp(min, max, x);
         // Then, map from the [0.0, 1.0] range to the [0.0, quantization_level] range, and quantize
-        // by casting to a usize.
+        // by casting to a u32.
         // TODO: should i round instead of cast?
-        lerp(0.0, quantization_level as f64, t) as usize
+        lerp(0.0, quantization_level as f64, t) as u32
     }
 }
